dockerdpeloydosyalar/requirements.txt
streamlit
pandas_gbq
google-cloud-bigquery
google-auth
google-api-python-client
google-cloud
google-auth-httplib2
numpy
pandas
plotly